refactor: eliminado módulo despacho y combinado con ventas
BREAKING CHANGE: es necesario destruir y recrear la base de datos para que siga funcionando el proyecto.
