+++ src/inventario/vistas_modelos/venta.py
@@ -4,12 +4,12 @@ from cuenta_usuario.restriccion import VistaRestringida
 from cuenta_usuario.enums.opciones import TipoUsuario
 from artemangaweb.mixins import MensajeResultadoFormMixin, TituloPaginaMixin
 from venta.models import Venta
-from venta.tipo_enum.estado_venta import EstadoVenta
+from venta.enums.opciones import EstadoVenta
 
 
 class VentasPendientesLisView(TituloPaginaMixin, VistaRestringida, ListView):
     titulo_pagina = "Ventas Pendientes"
-    usuarios_permitidos = [TipoUsuario.VENTAS]
+    usuarios_permitidos = [TipoUsuario.ADMINISTRADOR, TipoUsuario.VENTAS]
     model = Venta
     template_name = 'ventas/listado_ventas.html'
     queryset = Venta.objects.filter(estado=EstadoVenta.PENDIENTE.value)
@@ -20,7 +20,7 @@ class VentasPendientesLisView(TituloPaginaMixin, VistaRestringida, ListView):
 
 class VentasAprobadasLisView(TituloPaginaMixin, VistaRestringida, ListView):
     titulo_pagina = "Ventas Aprobadas"
-    usuarios_permitidos = [TipoUsuario.VENTAS]
+    usuarios_permitidos = [TipoUsuario.ADMINISTRADOR, TipoUsuario.VENTAS]
     model = Venta
     template_name = 'ventas/listado_ventas.html'
     queryset = Venta.objects.filter(estado=EstadoVenta.APROBADA.value)
@@ -30,7 +30,7 @@ class VentasAprobadasLisView(TituloPaginaMixin, VistaRestringida, ListView):
 
 
 class VentaUpdateView(VistaRestringida, MensajeResultadoFormMixin, UpdateView):
-    usuarios_permitidos = [TipoUsuario.VENTAS]
+    usuarios_permitidos = [TipoUsuario.ADMINISTRADOR, TipoUsuario.VENTAS]
     model = Venta
     template_name = 'CRUD/form_generico.html'
     success_url = reverse_lazy('ventas-validar')
